requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
urllib3==1.25.6 
requests==2.22.0
google-api-python-client==1.11.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
